# Deflate speed up 2