Improve deflate speed.
For the measurement, image data of 4 MB is used.

### Start LZ77 search from the end

|No.|before|after|
|:---:|---:|---:|
|1|19574 ms|3839 ms|
|2|19130 ms|3809 ms|
|3|19254 ms|3911 ms|
|4|19222 ms|3796 ms|
|5|19254 ms|3826 ms|